Wrote comments that make the code easy to understand , made readme file that explains all the tasks
