+++ controllers/priceConverterController.js
@@ -1,11 +1,10 @@
-// Filename: controllers/priceController.js
-import { calculatePrice } from "../services/priceService.js";
+import { calculatePrice } from "../services/priceConvertingService.js";
 
 export const getPrice = async (req, res) => {
   try {
-    const { fromCurrency, toCurrency, date } = req.body;
-    const price = await calculatePrice(fromCurrency, toCurrency, date);
-    res.json({ price });
+    const { fromcurrency, tocurrency, date } = req.body;
+    const price = await calculatePrice(fromcurrency, tocurrency, date);
+    res.status(200).json({ price });
   } catch (error) {
     console.error("Error:", error.message);
     res.status(500).json({ error: "Internal Server Error" });
