+++ models/Cryptocurrency.js
@@ -1,6 +1,5 @@
 import mongoose from 'mongoose';
 
-
 const CryptocurrencySchema = new mongoose.Schema({
   id: { type: String, required: true, unique: true },
   name: { type: String, required: true },
