+++ routes/companyListRoute.js
@@ -1,7 +1,5 @@
-// routes/companyRoutes.js
 import express from 'express';
 import { getCompanyList } from '../controllers/companyListController.js';
-
 
 const router = express.Router();
 router.post('/companies', getCompanyList);
