+++ services/priceConvertingService.js
@@ -1,8 +1,10 @@
-// File name: services/priceService.js
 import axios from "axios";
+import dotenv from "dotenv";
+dotenv.config();
 
-
-const converterFunction = async (fromCurrency, tocurrency, date) => {
+// This function fetch price of fromcurrency and toCurrency in USD on given date .Then on dividing
+// price of fromcurrency in USD with price of toCurrency in USD we get price of fromcurrency in terms of toCurrency.
+const converterFunction = async (fromCurrency, toCurrency, date) => {
   try {
     const apiKey = process.env.API_KEY
     const response1 = await axios.get(
@@ -16,7 +18,7 @@ const converterFunction = async (fromCurrency, tocurrency, date) => {
 
     const firstCurrInUsd= response1.data.market_data.current_price.usd;
     const response2 = await axios.get(
-      `https://api.coingecko.com/api/v3/coins/${tocurrency}/history?date=${date}&localization=false`,
+      `https://api.coingecko.com/api/v3/coins/${toCurrency}/history?date=${date}&localization=false`,
       {
         headers: {
         'Authorization':`Bearer ${apiKey}`
